Optimize code
